+++ backend/src/main/kotlin/estga/dadm/backend/repository/TreinoRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TreinoRepository : JpaRepository<Treino, Long> {
+interface TreinoRepository : JpaRepository<Treino, Int> {
     fun findByProfessorIdAndDiaSemanaOrderByHoraAsc(professorId: Int, diaSemana: String?): List<Treino>
 
     fun findByModalidadeIdInAndDiaSemanaOrderByHoraAsc(modalidadesIds: List<Int>, diaSemana: String?): List<Treino>
